+++ includes/class.llms.ajax.handler.php
@@ -2,7 +2,7 @@
 /**
  * LifterLMS AJAX Event Handler
  * @since    1.0.0
- * @version  3.14.2
+ * @version  [version]
  */
 if ( ! defined( 'ABSPATH' ) ) { exit; }
 
@@ -538,6 +538,7 @@ class LLMS_AJAX_Handler {
 
 		return array(
 			'html' => $html,
+			'question_id' => $question_id,
 		);
 
 	}
@@ -547,7 +548,7 @@ class LLMS_AJAX_Handler {
 	 * @param    [type]     $request  [description]
 	 * @return   [type]               [description]
 	 * @since    3.9.0
-	 * @version  3.9.0
+	 * @version  [version]
 	 */
 	public static function quiz_answer_question( $request ) {
 
@@ -581,7 +582,7 @@ class LLMS_AJAX_Handler {
 		$attempt->answer_question( $question_id, $answer );
 
 		// get the next question
-		$question_id = $attempt->get_next_question();
+		$question_id = $attempt->get_next_question( $question_id );
 
 		// return html for the next question
 		if ( $question_id ) {
@@ -594,6 +595,7 @@ class LLMS_AJAX_Handler {
 
 			return array(
 				'html' => $html,
+				'question_id' => $question_id,
 			);
 
 		} else {